+++ components/NfcShare.vue
@@ -11,16 +11,19 @@
       <template #default>
         <div class="row justify-between items-center">
           <!-- eslint-disable vue/no-v-html
-            this is ok to disable only because `contactlessSvg` has been
-            specifically sanitized -->
+          this is ok to disable only because `contactlessSvg` has been
+          specifically sanitized -->
           <div
-            style="height: 24px; margin-right: 8px;"
-            v-html="contactlessSvg" />
+          style="height: 24px; margin-right: 8px;"
+          v-html="contactlessSvg" />
           <!-- eslint-enable -->
           <div>
             Tap to Share
           </div>
         </div>
+        <q-tooltip>
+          Tap to Share is only available on Android devices.
+        </q-tooltip>
       </template>
     </q-btn>
   </div>