+++ components/ProfileForm.vue
@@ -103,7 +103,6 @@ export default {
     },
     profileOptions: {
       type: Array,
-      default: () => [],
       required: true
     }
   },